+++ src/httpErrorInterceptor/httpErrorInterceptor.ts
@@ -1,6 +1,6 @@
 import {Injectable, Inject, Optional, ClassProvider, InjectionToken} from '@angular/core';
 import {HttpInterceptor, HTTP_INTERCEPTORS, HttpEvent, HttpHandler, HttpErrorResponse, HttpRequest} from '@angular/common/http';
-import {Logger, LOGGER, IgnoredInterceptorsService, IgnoredInterceptorId, AdditionalInfo, Notifications} from '@anglr/common';
+import {Logger, LOGGER, IGNORED_INTERCEPTORS, Notifications} from '@anglr/common';
 import {isFunction, isArray} from '@jscrpt/common';
 import {Observable} from 'rxjs';
 import {tap} from 'rxjs/operators';
@@ -31,7 +31,6 @@ export class HttpErrorInterceptor implements HttpInterceptor
     constructor(@Optional() private _options: HttpErrorInterceptorOptions,
                 @Optional() private _internalServerErrorService: InternalServerErrorService,
                 @Optional() private _serverValidationService: ServerValidationService,
-                @Optional() private _ignoredInterceptorsService: IgnoredInterceptorsService,
                 @Optional() @Inject(ERROR_RESPONSE_MAP_PROVIDER) private _responseMapper: ResponseMapperFunction,
                 @Optional() @Inject(ERROR_HANDLING_NOTIFICATIONS) private _notifications: Notifications,
                 @Inject(LOGGER) private _logger: Logger)
@@ -54,14 +53,14 @@ export class HttpErrorInterceptor implements HttpInterceptor
      * @param req - Request to be intercepted
      * @param next - Next middleware that can be called for next processing
      */
-    public intercept(req: HttpRequest<any> & AdditionalInfo<IgnoredInterceptorId>, next: HttpHandler): Observable<HttpEvent<any>>
+    public intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>>
     {
         return next.handle(req)
             .pipe(tap(() => {}, (err: HttpErrorResponse) =>
             {
                 //client error, not response from server, or is ignored
                 if (err.error instanceof Error || 
-                    (this._ignoredInterceptorsService && this._ignoredInterceptorsService.isIgnored(HttpErrorInterceptor, req.additionalInfo)))
+                    req.context.get(IGNORED_INTERCEPTORS).some(itm => itm == HttpErrorInterceptor))
                 {
                     return;
                 }